pyproject.toml
[project]
name = "burnr8"
version = "0.4.0"
description = "Google Ads MCP server — manage your ad spend from Claude Code"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [
    { name = "Harrison Hesslink" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet",
    "Topic :: Office/Business",
]
dependencies = [
    "fastmcp>=2.0.0,<4.0.0",
    "google-ads>=25.1.0,<31.0.0",
    "python-dotenv>=1.0.0",
    "google-auth-oauthlib>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "ruff>=0.4.0"]

[project.urls]
Repository = "https://github.com/HarrisonHesslink/burnr8"
"Bug Tracker" = "https://github.com/HarrisonHesslink/burnr8/issues"
Changelog = "https://github.com/HarrisonHesslink/burnr8/blob/main/CHANGELOG.md"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
burnr8 = "burnr8.dashboard:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
burnr8 = ["py.typed"]

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]
ignore = [
    "E501",  # line length handled by formatter
    "SIM108",  # ternary operator — walrus pattern is fine
]

[tool.ruff.lint.isort]
known-first-party = ["burnr8"]
